lab2: update gains now that we have fixes for stuff

+++ labs/lab2/src/motor.h
@@ -5,9 +5,9 @@
 
 /* Types and constants */
 typedef struct {
-    int8_t kp;
-    int8_t ki;
-    int8_t kd;
+    int16_t kp;
+    int16_t ki;
+    int16_t kd;
 } pid_gains_t;
 
 
@@ -18,12 +18,12 @@ void motor_init();
 void reset_position();
 void set_setpoint(int16_t setpoint);
 uint16_t get_setpoint();
-uint16_t get_error();
-uint16_t get_torque();
+int32_t get_error();
+int32_t get_torque();
 void run_motor();
 uint8_t is_motor_running();
 
 int16_t update_pid(pid_gains_t * pid_gains);
-void update_gains(int8_t kp, int8_t kd, int8_t ki, pid_gains_t * gains);
+void update_gains(int16_t kp, int16_t kd, int16_t ki, pid_gains_t * gains);
 
 #endif
